2020/22, optimize
Made a bunch of poor choices up front thinking this would be a rotation
puzzle. We don't need a deque here, and it requires slower enumeration
than a list where we can use slices. Also flattened the data structures
to avoid creating copies everywhere.

The other surprising thing is that hashing a repr() versus tuples is way
slower. Huh.