2021/04, reconsidered
Trying to reduce iterations and squinting at this, it doesn't seem
necessary to iterate through each board for every ball. Since we know
the balls and what round they are drawn, we can run through every board
once, find the winning round for each, and then keep track of the first
and last winning board. This is faster.