2022/16, try to make saner
Use recursion to find the best path off a precomputed set of shortest
paths. For part 2, run the simulation twice and don't revisit the path
the first player used. This works fine for my input but still doesn't
work for the sample (probably because all valves can be opened in 26
minutes). Oh well, good enough.