2022/21, solve with sympy
A better approach to this problem is using an equation solver, which I
wasn't familiar with. This solution also has the benefit of passing for
both the sample input and my input which the binary search doesn't as I
didn't properly identify the boundary checks. I typically try to solve
these without imports, but this was a cool thing to learn, so including
it here separately.